Allow post checks to async functions
If functions are async, we should allow the post checks to run on the
results.  We do so by checking if the resulting entity is a promise.  If
that is the case, we run the solution on the awaited response.  We allow
overriding on what we should check in the "reason" argument, which can
now also be a javascript object.